backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
redis==5.0.1
cloudinary==1.38.0
python-dotenv==1.0.1
numpy==1.26.4
